requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
asgiref==3.3.4
    # via
    #   -r requirements/base.txt
    #   django
certifi==2020.12.5
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.14.5
    # via
    #   -r requirements/base.txt
    #   cryptography
chardet==4.0.0
    # via
    #   -r requirements/base.txt
    #   requests
coreapi==2.3.3
    # via
    #   -r requirements/base.txt
    #   drf-yasg
coreschema==0.0.4
    # via
    #   -r requirements/base.txt
    #   coreapi
    #   drf-yasg
cryptography==3.4.7
    # via
    #   -r requirements/base.txt
    #   pyjwt
defusedxml==0.7.1
    # via
    #   -r requirements/base.txt
    #   python3-openid
dj-rest-auth==2.1.4
    # via -r requirements/base.txt
django-admin-honeypot==1.1.0
    # via -r requirements/base.txt
django-allauth==0.44.0
    # via -r requirements/base.txt
django-anymail[sendgrid]==8.2
    # via -r requirements/base.txt
django-bootstrap3==15.0.0
    # via -r requirements/base.txt
django-debug-toolbar==3.2.1
    # via -r requirements/base.txt
django-environ==0.4.5
    # via -r requirements/base.txt
django-hijack==2.3.0
    # via -r requirements/base.txt
django-meta==2.0.0
    # via -r requirements/base.txt
django-phonenumber-field==5.1.0
    # via -r requirements/base.txt
django-phonenumbers==1.0.1
    # via -r requirements/base.txt
django==3.2
    # via
    #   -r requirements/base.txt
    #   dj-rest-auth
    #   django-allauth
    #   django-anymail
    #   django-bootstrap3
    #   django-debug-toolbar
    #   django-hijack
    #   django-phonenumber-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-jwt==1.11.0
    # via -r requirements/base.txt
djangorestframework-simplejwt==4.6.0
    # via -r requirements/base.txt
djangorestframework==3.12.4
    # via
    #   -r requirements/base.txt
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-yasg
drf-yasg==1.20.0
    # via -r requirements/base.txt
gitdb==4.0.7
    # via
    #   -r requirements/base.txt
    #   gitpython
gitpython==3.1.14
    # via -r requirements/base.txt
gunicorn==20.0.4
    # via -r requirements/dev.in
idna==2.10
    # via
    #   -r requirements/base.txt
    #   requests
importlib-metadata==4.0.0
    # via -r requirements/base.txt
inflection==0.5.1
    # via
    #   -r requirements/base.txt
    #   drf-yasg
itypes==1.2.0
    # via
    #   -r requirements/base.txt
    #   coreapi
jinja2==2.11.3
    # via
    #   -r requirements/base.txt
    #   coreschema
markupsafe==1.1.1
    # via
    #   -r requirements/base.txt
    #   jinja2
oauthlib==3.1.0
    # via
    #   -r requirements/base.txt
    #   requests-oauthlib
packaging==20.9
    # via
    #   -r requirements/base.txt
    #   drf-yasg
phonenumbers==8.12.21
    # via
    #   -r requirements/base.txt
    #   django-phonenumbers
pillow==8.2.0
    # via -r requirements/base.txt
psycopg2-binary==2.8.6
    # via -r requirements/dev.in
pycparser==2.20
    # via
    #   -r requirements/base.txt
    #   cffi
pyjwt[crypto]==1.7.1
    # via
    #   -r requirements/base.txt
    #   django-allauth
    #   djangorestframework-jwt
    #   djangorestframework-simplejwt
pyparsing==2.4.7
    # via
    #   -r requirements/base.txt
    #   packaging
python3-openid==3.2.0
    # via
    #   -r requirements/base.txt
    #   django-allauth
pytz==2021.1
    # via
    #   -r requirements/base.txt
    #   django
requests-oauthlib==1.3.0
    # via
    #   -r requirements/base.txt
    #   django-allauth
requests==2.25.1
    # via
    #   -r requirements/base.txt
    #   coreapi
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
ruamel.yaml.clib==0.2.2
    # via
    #   -r requirements/base.txt
    #   ruamel.yaml
ruamel.yaml==0.17.4
    # via
    #   -r requirements/base.txt
    #   drf-yasg
six==1.15.0
    # via
    #   -r requirements/base.txt
    #   django-meta
smmap==4.0.0
    # via
    #   -r requirements/base.txt
    #   gitdb
sqlparse==0.4.1
    # via
    #   -r requirements/base.txt
    #   django
    #   django-debug-toolbar
uritemplate==3.0.1
    # via
    #   -r requirements/base.txt
    #   coreapi
    #   drf-yasg
urllib3==1.26.4
    # via
    #   -r requirements/base.txt
    #   requests
wheel==0.36.2
    # via -r requirements/base.txt
zipp==3.4.1
    # via
    #   -r requirements/base.txt
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
